+++ config-server/src/main/kotlin/com/gitub/oopgurus/refactoringproblems/configserver/System.kt
@@ -16,39 +16,3 @@ class System(
 
     private fun notes(): String = entity.notes!!
 }
-
-
-interface WhatIWantToSystem {
-    fun id(id: Long)
-    fun on(on: Boolean)
-    fun notes(notes: String)
-}
-
-class SystemDtoBuilder: WhatIWantToSystem {
-    private lateinit var idGet: () -> Long
-    private lateinit var onGet: () -> Boolean
-    private lateinit var offGet: () -> Boolean
-    private lateinit var notesGet: () -> String
-
-    override fun id(id: Long) {
-        idGet = { id }
-    }
-
-    override fun on(on: Boolean) {
-        onGet = { on }
-        offGet = { !on }
-    }
-
-    override fun notes(notes: String) {
-        notesGet = { notes }
-    }
-
-    fun build(): SystemDto {
-        return SystemDto(
-            id = idGet(),
-            on = onGet(),
-            off =  onGet(),
-            notes = notesGet(),
-        )
-    }
-}
