+++ config-server/src/main/kotlin/com/gitub/oopgurus/refactoringproblems/configserver/WhatIWantToSystem.kt
@@ -1,0 +1,7 @@
+package com.gitub.oopgurus.refactoringproblems.configserver
+
+interface WhatIWantToSystem {
+    fun id(id: Long)
+    fun on(on: Boolean)
+    fun notes(notes: String)
+}
